% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdist.R
\name{plotdist}
\alias{plotdist}
\title{Plot Different Probability Distributions}
\usage{
plotdist(n, dist, ..., kde = FALSE)
}
\arguments{
\item{n}{number of sampled elements, n > 0}

\item{dist}{abbreviation for different distributions}

\item{...}{optional parameters for different distributions}

\item{kde}{draw a kernal density on top of the histogram}
}
\description{
Plot a histogram of n elements sampled from various distributions.
Possible Distributions:
\itemize{
  \item 'unif' -> Uniform(min, max)
  \item 'norm' -> Normal(mean, sd)
  \item 'chisq' -> Chi-Squared(df)
  \item 'exp' -> Exponential(rate)
}
}
\examples{
plotdist(50, 'unif', min = 10, max = 15)
plotdist(100, 'norm', mean = 5, sd = 2)
plotdist(500, 'chisq', df = 5)
plotdist(10, 'exp', rate = 3)
}
\author{
Matt Mackenzie
}
